+++ src/fetch_agent/fetch_agent/agent.py
@@ -1,5 +1,4 @@
 import asyncio
-from asyncio import sleep
 import datetime
 import socket
 import rclpy
@@ -9,19 +8,14 @@ from .minimal_publisher import MinimalPublisher
 from uagents import Agent, Context
 from uagents.setup import fund_agent_if_low
 
-from aca_protocols.station_register_protocol import (
-    StationRegisterRequest,
-)
 
 from aca_protocols.property_query_protocol import (
     PropertyData,
 )
 
-from aca_protocols.acs_registry_id import acs_id
 
-from .protocols.station import protocol as station_protocol
-from .protocols.car import protocol as car_protocol
-from .protocols.property_query import protocol as property_query_protocols
+from .communication.station import protocol as station_protocol, register_at_registry
+from .communication.car import protocol as car_protocol
 
 hostname = socket.gethostname()
 IPAddr = socket.gethostbyname(hostname)
@@ -35,7 +29,6 @@ agent = Agent(
 
 agent.include(station_protocol)
 agent.include(car_protocol)
-agent.include(property_query_protocols)
 
 rclpy.init()
 minimal_publisher = MinimalPublisher()
@@ -49,7 +42,7 @@ async def startup_event(ctx: Context):
     minimal_publisher.log_message(f"Agent: {agent.name} ({agent.address})")
 
     # run function in background so agent can fully start while registering
-    asyncio.ensure_future(register_at_registry(ctx))
+    asyncio.ensure_future(register_at_registry(ctx, agent))
 
     properties = PropertyData(
         open_time_frames=[(0, 0), (0, 0)],
@@ -61,31 +54,6 @@ async def startup_event(ctx: Context):
 
     ctx.storage.set("properties", properties.to_json())
     ctx.storage.set("expireAt", datetime.datetime.fromtimestamp(86400).timestamp())
-
-
-async def register_at_registry(ctx: Context):
-    while True:
-        if (
-            datetime.datetime.fromtimestamp(ctx.storage.get("expireAt"))
-            > datetime.datetime.now()
-        ):
-            await sleep(
-                ctx.storage.get("expireAt") - datetime.datetime.now().timestamp()
-            )
-            continue
-
-        ctx.logger.info(f"Trying to introduce: {agent.name} ({agent.address})")
-
-        properties = ctx.storage.get("properties")
-        properties = PropertyData.from_json(properties)
-        await ctx.send(
-            acs_id,
-            StationRegisterRequest(
-                lat=properties.geo_point[0], long=properties.geo_point[1]
-            ),
-        )
-
-        await sleep(5)
 
 
 def main():
