--- src/fetch_agent/fetch_agent/protocols/__init__.py
@@ -1,3 +1,0 @@
-from .car import protocol as protocol
-from .property_query import protocol as protocol
-from .station import protocol as protocol
